nvidia: 358.16 -> 364.12

+++ pkgs/os-specific/linux/nvidia-x11/builder.sh
@@ -62,7 +62,9 @@ installPhase() {
 
       # nvidia's EGL stack seems to expect libGLESv2.so.2 to be available
       if [ $(basename "$libname_short") == "libGLESv2.so" ]; then
-          ln -srnf "$libname" "$libname_short.2"
+          if [ ! -e "$libname_short.2" ]; then
+            ln -srnf "$libname" "$libname_short.2"
+          fi
       fi
 
       if [[ "$libname" != "$libname_short" ]]; then
@@ -117,10 +119,10 @@ installPhase() {
     fi
 
     # For simplicity and dependency reduction, don't support the gtk3 interface.
-    rm $out/lib/libnvidia-gtk3.*
+    rm -fv $out/lib/libnvidia-gtk3.*
 
     # We distribute these separately in `libvdpau`
-    rm "$out"/lib/libvdpau{.*,_trace.*}
+    rm -fv "$out"/lib/libvdpau{.*,_trace.*}
 
     # Move VDPAU libraries to their place
     mkdir "$out"/lib/vdpau
